automation/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.3
Django==5.2.2
et_xmlfile==2.0.0
idna==3.10
kombu==5.5.4
msal==1.32.3
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
google-generativeai
django_celery_beat